more rebasing and refactoring...
