+++ src/main/java/com/bkj/search/GomezClasses/FileManager.java
@@ -7,7 +7,7 @@
  *
  * @author Kelvin
  */
-package GomezClasses;
+package com.bkj.search.GomezClasses;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
