+++ src/main/java/com/bkj/search/JonsGUI/File_Add.java
@@ -1,4 +1,4 @@
-package GUI;
+package com.bkj.search.JonsGUI;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
