+++ src/main/java/com/bkj/search/JonsGUI/Inverted_Index.java
@@ -1,4 +1,4 @@
-package GUI;
+package com.bkj.search.JonsGUI;
 
 import java.util.TreeMap;
