+++ src/main/java/com/bkj/search/JonsGUI/Main_UI.java
@@ -1,4 +1,4 @@
-package GUI;
+package com.bkj.search.JonsGUI;
 
 import javax.swing.JOptionPane;
