+++ src/main/java/com/bkj/search/bradGUI/AdministrationWindow.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.bkj.search.gui.AdministrationWindow">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.bkj.search.bradGUI.AdministrationWindow">
   <grid id="27dc6" binding="topPanel" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="20" y="20" width="500" height="400"/>
