+++ src/main/java/com/bkj/search/bradGUI/NotImplementedDialog.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.bkj.search.gui.NotImplementedDialog">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.bkj.search.bradGUI.NotImplementedDialog">
   <grid id="cbd77" binding="contentPanel" layout-manager="BorderLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="48" y="54" width="436" height="297"/>
