+++ src/main/java/com/bkj/search/bradUtils/IFileInvertedIndex.java
@@ -1,4 +1,4 @@
-package com.bkj.search.utils;
+package com.bkj.search.bradUtils;
 
 import java.io.IOException;
 import java.util.Date;
