+++ src/main/java/com/bkj/search/bradUtils/Loadable.java
@@ -1,6 +1,4 @@
-package com.bkj.search.utils;
-
-import com.bkj.search.gui.MainWindow;
+package com.bkj.search.bradUtils;
 
 public interface Loadable<C> {
     void loadFromJson(C b);
