+++ src/main/java/com/bkj/search/bradUtils/MD5Checksum.java
@@ -1,8 +1,6 @@
-package com.bkj.search.utils;
+package com.bkj.search.bradUtils;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 
 /**
