+++ src/main/java/com/bkj/search/bradUtils/SearchUIModel.java
@@ -1,4 +1,4 @@
-package com.bkj.search.utils;
+package com.bkj.search.bradUtils;
 
 import com.google.gson.Gson;
 
@@ -7,7 +7,6 @@ import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.*;
-import java.util.function.Function;
 
 public class SearchUIModel implements IFileStore, Saveable {
